school_lessons/03.image-clusterisation/requirements.txt
sklearn
Pillow
numpy
